Refactor CommandProcessor to InteractionProcessor (more global) + added Button support
