--- src/main/java/joseta/annotations/modules/CommandModule.java
@@ -1,11 +1,0 @@
-package joseta.annotations.modules;
-
-import net.dv8tion.jda.api.events.interaction.command.*;
-
-public abstract class CommandModule {
-    protected SlashCommandInteractionEvent event;
-
-    public CommandModule(SlashCommandInteractionEvent event) {
-        this.event = event;
-    }
-}
