+++ src/main/java/joseta/annotations/types/ButtonInteraction.java
@@ -1,0 +1,9 @@
+package joseta.annotations.types;
+
+import java.lang.annotation.*;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface ButtonInteraction {
+    String id() default "";
+}
