+++ src/main/java/joseta/commands/MiscCommands.java
@@ -1,21 +1,20 @@
 package joseta.commands;
 
-import joseta.annotations.modules.*;
+import joseta.annotations.*;
 import joseta.annotations.types.*;
 import net.dv8tion.jda.api.entities.*;
 import net.dv8tion.jda.api.events.interaction.command.*;
 
-public class MiscCommands extends CommandModule {
+@InteractionModule
+public class MiscCommands {
 
-    public MiscCommands(SlashCommandInteractionEvent event) { super(event); }
-
-    @SlashCommand(description = "Check the bot's responsiveness")
-    public void ping() {
+    @SlashCommandInteraction(description = "Check the bot's responsiveness")
+    public void ping(SlashCommandInteractionEvent event) {
         event.reply("Poong! "+ event.getJDA().getGatewayPing() +"ms " + event.getName()).queue();
     }
 
-    @SlashCommand(description = "Responds with ping")
-    public void pong(@Option User user) {
+    @SlashCommandInteraction(description = "Responds with ping")
+    public void pong(SlashCommandInteractionEvent event, @Option User user) {
         if (user == null) {
             event.reply("l4p1n was here").queue();
             return;
