+++ src/main/resources/bundles/bundle_en.properties
@@ -1,0 +1,2 @@
+ping.name = pingteaaa
+ping.description = A cute animal that hops.
