+++ src/main/resources/bundles/bundle_fr.properties
@@ -1,0 +1,2 @@
+ping.name = lapin
+ping.description = Un animal mignon qui saute.
